#21 Rename page.ts to createGetStaticPaths.ts
