+++ lib/createGetStaticPaths.ts
