+++ pages/tags/[tag].tsx
@@ -3,7 +3,7 @@ import { getAllTagsByDir, getNotePreviewsByDirAndTag, NotePreview, Tag } from '.
 import NoteListView from '../../components/NoteListView';
 import { getPageSize } from '../../lib/configuration';
 import TagListHeader from '../../components/TagListHeader';
-import createGetStaticPaths from '../../lib/page';
+import createGetStaticPaths from '../../lib/createGetStaticPaths';
 
 export default function TagPage(
   { tags, tag, articles, pageNumber, isFirstPage, isLastPage }: TagPageProps,
